a little modify
