+++ src/main/java/llcweb/com/service/ImageService.java
@@ -1,6 +1,6 @@
 package llcweb.com.service;
 
-import llcweb.com.domain.UsefulImage;
+import llcweb.com.domain.entity.UsefulImage;
 import llcweb.com.domain.models.Image;
 import org.springframework.data.domain.Page;
